Only link to ssmi and radar files if they exist and are non-zero (test -s)
git-svn-id: https://svn-wrf-var.cgd.ucar.edu/branches/afwa_2.2@3053 70815b0f-b814-0410-81f7-dfa10a885da6

+++ scripts/da_run_wrfvar.ksh
@@ -283,8 +283,12 @@ echo "WINDOW_END            $WINDOW_END"
       fi
    else
       ln -fs $OB_DIR/${DATE}/ob.ascii ob01.ascii
-      ln -fs $OB_DIR/${DATE}/ob.ssmi  ob01.ssmi
-      ln -fs $OB_DIR/${DATE}/ob.radar ob01.radar
+      if [[ -s $OB_DIR/${DATE}/ob.ssmi ]]; then
+         ln -fs $OB_DIR/${DATE}/ob.ssmi ob01.ssmi
+      fi
+      if [[ -s $OB_DIR/${DATE}/ob.radar ]]; then
+         ln -fs $OB_DIR/${DATE}/ob.radar ob01.radar
+      fi
    fi
 
    for FILE in $OB_DIR/$DATE/*.bufr; do
